Update sequence numbers and add new Armstrong number tests
Sequence numbers have been updated in 'A seven digit number that is an Armstrong number.bru', 'Error when number is not specified.bru' and 'A seven digit number that is not an Armstrong number.bru' files. Besides, new Armstrong number test scenarios have been introduced for three-digit and four-digit numbers. These changes will help refine test flow and improve the overall test coverage of the Armstrong numbers' functionality in the API.

@coderabbitai ignore